requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
gunicorn>=20.1,<21.0
psycopg2>=2.9,<3.0
celery>=5.2,<6.0
redis>=4.0,<5.0
python-decouple>=3.5,<4.0
django-cors-headers>=3.7,<4.0
django-environ>=0.4,<1.0
django-allauth>=0.47,<1.0
whitenoise>=5.2,<6.0
django-extensions>=3.1,<4.0
django-crispy-forms>=1.12,<2.0
requests>=2.25,<3.0
mysqlclient>=2.0,<3.0
sshtunnel>=0.4,<1.0
django-filter>=22.0,<23.0
boto3>=1.18,<2.0
